Centralize Flow types
