+++ .flowconfig
@@ -1,6 +1,7 @@
 [ignore]
 .*/react-cosmos-voyager/src/use-cases/.*
 💩 Ignore packages with type issues
+.*/node_modules/graphql/.*
 .*/node_modules/apollo-client/.*
 .*/node_modules/apollo-link-core/.*
 .*/node_modules/react-apollo/.*
