+++ package.json
@@ -55,7 +55,7 @@
     "eslint-plugin-flowtype": "^2.46.1",
     "eslint-plugin-react": "^7.7.0",
     "file-loader": "^1.1.11",
-    "flow-bin": "^0.68.0",
+    "flow-bin": "^0.69.0",
     "flow-typed": "^2.3.0",
     "glob": "^7.1.2",
     "html-webpack-plugin": "^3.0.7",
