+++ packages/react-cosmos-flow/linked-list.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type LinkedItem<Item> = {
+  value: Item,
+  next: () => LinkedItem<Item>
+};
