+++ packages/react-cosmos-flow/loader.js
@@ -1,15 +1,5 @@
 // @flow
 
-type ExcludePattern = string | RegExp;
-
-export type ExcludePatterns = ExcludePattern | Array<ExcludePattern>;
-
-export type PlaygroundOpts = {
-  loaderUri: string,
-  projectKey: string,
-  webpackConfigType: 'default' | 'custom'
-};
-
 export type LoaderOpts = {
   containerQuerySelector?: string
 };
