+++ packages/react-cosmos-flow/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "react-cosmos-flow",
+  "version": "4.0.0-beta.5",
+  "description": "Flow types and utilities",
+  "repository": "https://github.com/react-cosmos/react-cosmos/tree/master/packages/react-cosmos-flow",
+  "license": "MIT",
+  "dependencies": {},
+  "xo": false
+}
