+++ packages/react-cosmos-flow/playground.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type PlaygroundOpts = {
+  loaderUri: string,
+  projectKey: string,
+  webpackConfigType: 'default' | 'custom'
+};
