+++ packages/react-cosmos-flow/react.js
@@ -1,0 +1,7 @@
+// @flow
+
+import { Component } from 'react';
+
+import type { ElementRef } from 'react';
+
+export type ComponentRef = ElementRef<typeof Component>;
