+++ packages/react-cosmos-loader/src/create-context.js
@@ -6,12 +6,12 @@ import createStateProxy from 'react-cosmos-state-proxy';
 import Loader from './components/Loader';
 import { isComponentClass } from './utils/is-component-class';
 
+import type { ComponentRef } from 'react-cosmos-flow/react';
 import type {
-  ComponentRef,
   Wrapper,
   ContextArgs,
   ContextFunctions
-} from './types';
+} from 'react-cosmos-flow/context';
 
 let wrapper: ?Wrapper;
 const StateProxy = createStateProxy({ updateInterval: 50 });
