+++ packages/react-cosmos-playground/src/components/screens/NoLoaderScreen/index.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react';
 import DisplayScreen from '../DisplayScreen';
 import style from '../DisplayScreen/index.less';
 
-import type { PlaygroundOpts } from 'react-cosmos-shared/src/types';
+import type { PlaygroundOpts } from 'react-cosmos-flow/playground';
 
 type Props = {
   options: PlaygroundOpts
