+++ packages/react-cosmos-scripts/src/upgrade-fixtures.js
@@ -8,7 +8,7 @@ import { getCosmosConfig } from 'react-cosmos-config';
 import getFilePaths from 'react-cosmos-voyager';
 import { addComponentToFixture } from './transforms/add-component-to-fixture';
 
-import type { Config } from 'react-cosmos-config/src';
+import type { Config } from 'react-cosmos-flow/config';
 
 const { keys } = Object;
