--- packages/react-cosmos-shared/src/jest/index.js
@@ -1,7 +1,0 @@
-// @flow
-
-import type { JestMock } from './types';
-
-export function getMock(fn: any): JestMock {
-  return fn.mock;
-}
