+++ packages/react-cosmos/src/server/embed-modules-webpack-loader.js
@@ -6,8 +6,8 @@ import { moduleExists } from 'react-cosmos-shared/lib/server';
 import getFilePaths from 'react-cosmos-voyager';
 import { findFixtureFiles } from 'react-cosmos-voyager2/lib/server';
 
-import type { Config } from 'react-cosmos-config/src';
-import type { FixtureFile } from 'react-cosmos-voyager2/src/types';
+import type { Config } from 'react-cosmos-flow/config';
+import type { FixtureFile } from 'react-cosmos-flow/module';
 
 const { keys } = Object;
